后台商城管理，调整数据存储格式，增加商品生命周期 接入ping++支付，微信扫码支付

+++ service/src/main/java/com/bestomb/service/IPingggService.java
@@ -18,7 +18,7 @@ public interface IPingggService {
      * @return
      * @throws Exception
      */
-    String getCharge(int amount, String channel, Integer memberId, HttpServletRequest request) throws Exception;
+    String getCharge(double amount, String channel, Integer memberId, HttpServletRequest request) throws Exception;
 
     /**
      * 接受、验证Webhooks并且更新会员充值支付信息
