+++ web/src/main/java/com/bestomb/controller/api/PingppController.java
@@ -34,7 +34,7 @@ public class PingppController {
      */
     @RequestMapping("/getCharge")
     @ResponseBody
-    public String getCharge(int amount, String channel, HttpServletRequest request) throws Exception {
+    public String getCharge(double amount, String channel, HttpServletRequest request) throws Exception {
         /**
          * 从session池中获取系统用户信息
          */
